remove pending changes notification component (to complexe) and make notification apear also for browser navigation
